gui-backend/requirements.txt
fastapi==0.81.0
pydantic==1.10.1
uvicorn==0.18.3
pymongo==4.2.0
pandas==1.4.4
orjson==3.8.0
python-dotenv==0.21.0
